pyproject.toml
[project]
name = "topo-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "dotenv>=0.9.9",
    "folium>=0.20.0",
    "kaleido>=1.2.0",
    "mapbox-vector-tile>=2.2.0",
    "mercantile>=1.2.1",
    "notebook>=7.5.0",
    "numpy>=2.3.5",
    "osdatahub>=1.3.2",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.25.3",
]
